feat: finishing project

+++ src/main/java/br/com/sobreiraromulo/ordersms/listener/dto/OrderItemEvent.java
@@ -3,6 +3,6 @@ package br.com.sobreiraromulo.ordersms.listener.dto;
 import java.math.BigDecimal;
 
 public record OrderItemEvent(String produto,
-                             Long quantidade,
+                             Integer quantidade,
                              BigDecimal preco) {
 }
